+++ src/molang/syntax/parse.ts
@@ -1,5 +1,4 @@
 import { OffsetWord } from "bc-minecraft-bedrock-types/lib/types";
-import { Token, TokenType, tokenize } from "./tokens";
 import {
   ArrayAccessNode,
   AssignmentNode,
@@ -12,26 +11,15 @@ import {
   NodeType,
   NullishCoalescingNode,
   ResourceReferenceNode,
-  StatementSequenceNode,
   StringLiteralNode,
   UnaryOperationNode,
   VariableNode,
 } from "./nodes";
-
-/** Represents a syntax error in the Molang code */
-export class MolangSyntaxError extends Error {
-  constructor(message: string, public position: number, public code: string) {
-    super(message);
-    this.name = "MolangSyntaxError";
-  }
-
-  static fromToken(token: Token, message: string) {
-    return new MolangSyntaxError(message, token.position, token.value);
-  }
-  static fromTokens(tokens: Token[], message: string) {
-    return new MolangSyntaxError(message, tokens[0].position, tokens.map((i) => i.value).join(""));
-  }
-}
+import { Token, TokenType, tokenize } from "./tokens";
+import { getMatchingTokenSlice } from "./util";
+import { SyntaxBuilder } from "./builder";
+import { MolangSyntaxError } from "./errors";
+import { processOperators } from "./operators";
 
 /** Main function to parse Molang code into a syntax tree */
 export function parseMolang(line: OffsetWord): ExpressionNode[] {
@@ -45,39 +33,11 @@ export function parseMolang(line: OffsetWord): ExpressionNode[] {
     .map(parseTokens);
 }
 
-class SyntaxBuilder {
-  result: StatementSequenceNode;
-
-  constructor(position: number) {
-    this.result = {
-      type: NodeType.StatementSequence,
-      statements: [],
-      position: position,
-    };
-  }
-
-  add<T extends ExpressionNode>(node: T): T {
-    if (node) this.result.statements.push(node);
-    return node;
-  }
-  remove<T extends ExpressionNode>(node: T) {
-    this.result.statements = this.result.statements.filter((item) => item !== node);
-  }
-  replace<T extends ExpressionNode, U extends ExpressionNode>(original: T, newnode: U): U {
-    this.result.statements.forEach((item, index, nodes) => {
-      if (item === original) {
-        nodes[index] = newnode;
-      }
-    });
-    return newnode;
-  }
-  build(): ExpressionNode {
-    if (this.result.statements.length === 1) return this.result.statements[0];
-
-    return this.result;
-  }
-}
-
+/**
+ * Converts the given tokens into nodes
+ * @param tokens
+ * @returns
+ */
 function parseTokens(tokens: Token[]) {
   tokens = trimBraces(tokens);
   tokens = trimEnding(tokens);
@@ -141,222 +101,6 @@ function parseTokens(tokens: Token[]) {
   processOperators(builder);
 
   return builder.build();
-}
-
-function processOperators(builder: SyntaxBuilder) {
-  processNullishCoalescing(builder); // ??
-
-  // Process unary operators (highest precedence)
-  processUnaryOperators(builder); // ? or -u
-
-  processBinaryOperators(builder, "||");
-  processBinaryOperators(builder, "&&");
-  processBinaryOperators(builder, "==");
-  processBinaryOperators(builder, "!=");
-  processBinaryOperators(builder, "<");
-  processBinaryOperators(builder, "<=");
-  processBinaryOperators(builder, ">");
-  processBinaryOperators(builder, ">=");
-  processBinaryOperators(builder, "+");
-  processBinaryOperators(builder, "-");
-  processBinaryOperators(builder, "*");
-  processBinaryOperators(builder, "/");
-  processBinaryOperators(builder, "%");
-
-  processTernaryOperators(builder); // <cond> ? <true> : <false>
-
-  // Process assignments last (right-to-left associativity)
-  processAssignments(builder); // =
-}
-
-function processNullishCoalescing(builder: SyntaxBuilder) {
-  const statements = builder.result.statements;
-
-  for (let i = 0; i < statements.length; i++) {
-    const current = statements[i];
-
-    if (current.type !== NodeType.NullishCoalescing) continue;
-    // Find left operand
-    if (i === 0) {
-      throw new MolangSyntaxError("Nullish coalescing operator missing left operand", current.position, "??");
-    }
-    // Find right operand
-    if (i === statements.length - 1) {
-      throw new MolangSyntaxError("Nullish coalescing operator missing right operand", current.position, "??");
-    }
-
-    // Update the nullish coalescing node
-    current.left = statements[i - 1];
-    current.right = statements[i + 1];
-
-    // Remove the operands from the statements array
-    builder.result.statements.splice(i - 1, 3, current);
-
-    // Adjust index since we removed elements
-    i -= 1;
-  }
-}
-
-function processBinaryOperators(builder: SyntaxBuilder, operator: string) {
-  const statements = builder.result.statements;
-
-  for (let i = 0; i < statements.length; i++) {
-    const current = statements[i];
-
-    if (current.type === NodeType.BinaryOperation && current.operator === operator) {
-      // Find left operand
-      if (i === 0) {
-        throw new MolangSyntaxError(`Binary operator '${operator}' missing left operand`, current.position, operator);
-      }
-      // Find right operand
-      if (i === statements.length - 1) {
-        throw new MolangSyntaxError(`Binary operator '${operator}' missing right operand`, current.position, operator);
-      }
-
-      // Update the binary operation node
-      current.left = statements[i - 1];
-      current.right = statements[i + 1];
-
-      // Remove the operands from the statements array
-      builder.result.statements.splice(i - 1, 3, current);
-
-      // Adjust index since we removed elements
-      i -= 1;
-    }
-  }
-}
-
-function processTernaryOperators(builder: SyntaxBuilder) {
-  const statements = builder.result.statements;
-
-  for (let i = 0; i < statements.length; i++) {
-    const current = statements[i];
-    if (current.type !== NodeType.Conditional) continue;
-
-    // Find condition (left operand)
-    if (i === 0) {
-      throw new MolangSyntaxError("Ternary operator missing condition", current.position, "?");
-    }
-
-    const condition = statements[i - 1];
-
-    // Find the colon marker and expressions
-    let colonIndex = -1;
-    let trueExpr: ExpressionNode | null = null;
-    let falseExpr: ExpressionNode | null = null;
-
-    // Look for the pattern: condition ? trueExpr : falseExpr
-    for (let j = i + 1; j < statements.length; j++) {
-      const stmt = statements[j];
-      if (stmt.type === NodeType.Marker && (stmt as MarkerNode).token.type === TokenType.Colon) {
-        colonIndex = j;
-        break;
-      }
-    }
-
-    if (colonIndex === -1) {
-      throw new MolangSyntaxError("Ternary operator missing colon", current.position, "?");
-    }
-
-    // Extract true and false expressions
-    if (i + 1 < colonIndex) {
-      // If there are statements between ? and :, they form the true expression
-      if (colonIndex - i - 1 === 1) {
-        trueExpr = statements[i + 1];
-      } else {
-        // Multiple statements - create a sequence
-        trueExpr = {
-          type: NodeType.StatementSequence,
-          statements: statements.slice(i + 1, colonIndex),
-          position: statements[i + 1].position,
-        } as StatementSequenceNode;
-      }
-    }
-
-    if (colonIndex + 1 < statements.length) {
-      falseExpr = statements[colonIndex + 1];
-    }
-
-    if (!trueExpr) {
-      throw new MolangSyntaxError("Ternary operator missing true expression", current.position, "?");
-    }
-
-    if (!falseExpr) {
-      throw new MolangSyntaxError("Ternary operator missing false expression", current.position, "?");
-    }
-
-    // Update the conditional expression node
-    (current as ConditionalExpressionNode).condition = condition;
-    (current as ConditionalExpressionNode).trueExpression = trueExpr;
-    (current as ConditionalExpressionNode).falseExpression = falseExpr;
-
-    // Remove all the processed elements and replace with the complete ternary
-    const elementsToRemove = colonIndex - i + 2; // condition + ? + trueExpr + : + falseExpr
-    builder.result.statements.splice(i - 1, elementsToRemove, current);
-
-    // Adjust index
-    i -= 1;
-  }
-}
-
-function processAssignments(builder: SyntaxBuilder) {
-  const statements = builder.result.statements;
-
-  if (statements.length === 3 && statements[1].type === NodeType.Assignment) {
-    // Update the assignment node
-    const n = statements[1];
-    n.left = statements[0];
-    n.right = statements[2];
-    builder.result.statements = [n];
-    return;
-  }
-
-  // Process right-to-left for right associativity
-  for (let i = statements.length - 1; i >= 0; i--) {
-    const current = statements[i];
-
-    if (current.type !== NodeType.Assignment) continue;
-    // Find left operand
-    if (i === 0) {
-      throw new MolangSyntaxError("Assignment operator missing left operand", current.position, "=");
-    }
-    // Find right operand
-    if (i === statements.length - 1) {
-      throw new MolangSyntaxError("Assignment operator missing right operand", current.position, "=");
-    }
-
-    // Update the assignment node
-    current.left = statements[i - 1];
-    current.right = statements[i + 1];
-
-    // Remove the operands from the statements array
-    builder.result.statements.splice(i - 1, 3, current);
-  }
-}
-
-/**
- * ? and -u
- */
-function processUnaryOperators(builder: SyntaxBuilder) {
-  const statements = builder.result.statements;
-
-  for (let i = 0; i < statements.length; i++) {
-    const current = statements[i];
-    if (current.type !== NodeType.UnaryOperation) continue;
-    // Find operand (should be to the right for prefix operators)
-    if (i === statements.length - 1) {
-      throw new MolangSyntaxError(
-        `Unary operator '${current.operator}' missing operand`,
-        current.position,
-        current.operator
-      );
-    }
-    // Update the unary operation node
-    current.operand = statements[i + 1];
-
-    // Remove the operand from the statements array
-    builder.result.statements.splice(i + 1, 1);
-  }
 }
 
 /** Filter () {} [] from start or finish if they match */
@@ -552,61 +296,4 @@ function splitTokens(tokens: Token[], predicate: (item: Token) => boolean): Arra
   }
 
   return result;
-}
-
-function getMatchingTokenSlice(tokens: Token[], startIndex: number): Token[] {
-  if (startIndex >= tokens.length) {
-    throw new Error("Start index is out of bounds");
-  }
-
-  const startToken = tokens[startIndex];
-  let targetType: TokenType;
-  let counterType: TokenType;
-
-  // Determine what we're looking for based on the starting token
-  switch (startToken.type) {
-    case TokenType.OpenBrace:
-      targetType = TokenType.CloseBrace;
-      counterType = TokenType.OpenBrace;
-      break;
-    case TokenType.OpenParen:
-      targetType = TokenType.CloseParen;
-      counterType = TokenType.OpenParen;
-      break;
-    case TokenType.OpenBracket:
-      targetType = TokenType.CloseBracket;
-      counterType = TokenType.OpenBracket;
-      break;
-    default:
-      throw new Error(`Token at index ${startIndex} is not an opening bracket, brace, or parenthesis`);
-  }
-
-  let level = 1; // We start with level 1 since we found the opening token
-  let endIndex = -1;
-
-  // Search for the matching closing token
-  for (let i = startIndex + 1; i < tokens.length; i++) {
-    const token = tokens[i];
-
-    if (token.type === counterType) {
-      level++;
-    } else if (token.type === targetType) {
-      level--;
-      if (level === 0) {
-        endIndex = i;
-        break;
-      }
-    }
-  }
-
-  // Validate that we found the matching closing token
-  if (endIndex === -1) {
-    const tokenName =
-      startToken.type === TokenType.OpenBrace ? "{" : startToken.type === TokenType.OpenParen ? "(" : "[";
-    const closingName = targetType === TokenType.CloseBrace ? "}" : targetType === TokenType.CloseParen ? ")" : "]";
-    throw new Error(`Couldn't find the matching closing ${closingName} for ${tokenName} at index ${startIndex}`);
-  }
-
-  // Return the slice including both opening and closing tokens
-  return tokens.slice(startIndex, endIndex + 1);
 }
