+++ src/molang/syntax/tokens.ts
@@ -39,21 +39,8 @@ export enum TokenType {
 }
 
 function isDigit(ch: string): boolean {
-  switch (ch) {
-    case "0":
-    case "1":
-    case "2":
-    case "3":
-    case "4":
-    case "5":
-    case "6":
-    case "7":
-    case "8":
-    case "9":
-      return true;
-    default:
-      return false;
-  }
+  const code = ch.charCodeAt(0);
+  return code >= 48 && code <= 57; // 0-9
 }
 
 function isAlpha(ch: string): boolean {
